refactor: only allow secondary imports
- #1461
