+++ src/app/app.component.ts
@@ -2,7 +2,8 @@ import { Component, OnInit, Renderer2, ElementRef } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
 import { filter } from 'rxjs/operators';
 import { VERSION as VERSION_ALAIN, TitleService } from '@delon/theme';
-import { VERSION as VERSION_ZORRO, NzModalService } from 'ng-zorro-antd';
+import { VERSION as VERSION_ZORRO } from 'ng-zorro-antd/version';
+import { NzModalService } from 'ng-zorro-antd/modal';
 
 @Component({
   selector: 'app-root',
