+++ src/app/layout/default/setting-drawer/setting-drawer.component.ts
@@ -1,6 +1,6 @@
 import { Component, ChangeDetectionStrategy, NgZone, Inject, ChangeDetectorRef } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
-import { NzMessageService } from 'ng-zorro-antd';
+import { NzMessageService } from 'ng-zorro-antd/message';
 import { LazyService, copy, deepCopy } from '@delon/util';
 import { SettingsService } from '@delon/theme';
