+++ src/app/routes/dashboard/monitor/monitor.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
-import { NzMessageService } from 'ng-zorro-antd';
+import { NzMessageService } from 'ng-zorro-antd/message';
 import { zip } from 'rxjs';
 import { _HttpClient } from '@delon/theme';
