+++ src/app/routes/passport/register-result/register-result.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { NzMessageService } from 'ng-zorro-antd';
+import { NzMessageService } from 'ng-zorro-antd/message';
 import { ActivatedRoute } from '@angular/router';
 
 @Component({
