+++ src/app/routes/pro/list/basic-list/basic-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
-import { NzMessageService } from 'ng-zorro-antd';
+import { NzMessageService } from 'ng-zorro-antd/message';
 import { _HttpClient, ModalHelper } from '@delon/theme';
 import { ProBasicListEditComponent } from './edit/edit.component';
