+++ src/app/routes/pro/profile/basic/basic.component.ts
@@ -1,5 +1,5 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { NzMessageService } from 'ng-zorro-antd';
+import { NzMessageService } from 'ng-zorro-antd/message';
 import { _HttpClient } from '@delon/theme';
 import { STColumn } from '@delon/abc';
 import { tap } from 'rxjs/operators';
