+++ src/app/routes/pro/result/success/success.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { NzMessageService } from 'ng-zorro-antd';
+import { NzMessageService } from 'ng-zorro-antd/message';
 
 @Component({
   selector: 'app-result-success',
